refactor(test): 테스트 클래스 패키지 구조 도메인과 동일하게 수정
