+++ src/test/java/kitchenpos/product/ui/ProductAcceptanceTestFixture.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui;
+package kitchenpos.product.ui;
 
 import static org.assertj.core.api.Assertions.assertThat;
